Initial implementation for reading from serial port
This is a small program to begin with. It will open a serial device,
which in this case is a bluetooth device MIMU22BT connected over serial.

This program writes an initial command word into MIMU22BT so that the
device starts sending data over bluetooth. It then keeps reading data
byte by byte and writes into a file.